Fix final 5 LOW findings: remove all thin wrappers, dead code, and smells
Thin wrappers replaced with aliases:
- parser.py: _collect_json_ld_nodes = utils.flatten_json_ld (was 6-line function)
- signals_audit.py: _flatten_json_ld = utils.flatten_json_ld (was 6-line function)
- scraper.py: _now_iso = utils.now_iso (was 5-line function)
- run_pre_crawl_analysis.py: _now_iso = utils.now_iso (was 2-line function)
- run_pre_crawl_analysis.py: _sanitise = utils.sanitise_csv_value (was 5-line function)

Dead code removed:
- storage.py: removed save_project() — defined but never called from any module
- gui.py: removed inline 'import itertools' (moved to top-level)
- utils.py: removed inline 'import os' (already at top-level)

Performance fix:
- run_pre_crawl_analysis.py: eliminated double urlparse() in _pick_diverse

Type annotation fix:
- gui.py: _human_size(nbytes: float) — was int but gets divided by 1024

Config fix:
- sitemap.py: _fetch_xml accepts user_agent and timeout parameters instead
  of reading module-level globals directly

Import fix:
- run_background_crawl.py: removed duplicate 'import storage', use
  storage_module consistently

All 225 tests pass (71 unit + 113 Playwright + 41 real data).

+++ gui.py
@@ -64,6 +64,7 @@ from __future__ import annotations
 
 import csv
 import io
+import itertools
 import json
 import logging
 import os
@@ -389,7 +390,7 @@ def _start_crawl_thread(
 
 # ── CSV / metrics helpers ────────────────────────────────────────────────
 
-def _human_size(nbytes: int) -> str:
+def _human_size(nbytes: float) -> str:
     """Format a byte count as a human-readable string (e.g. ``4.2 MB``)."""
     for unit in ("B", "KB", "MB", "GB"):
         if nbytes < 1024:
@@ -471,7 +472,6 @@ def _read_csv_page(
         A 4-tuple of (headers, rows_on_page, total_rows, total_pages).
         All values are empty/zero when the file does not exist.
     """
-    import itertools
     if not os.path.isfile(filepath):
         return [], [], 0, 0
